placed process image function in enum
