+++ src/components/organization/EditOrgdetailsModal.tsx
@@ -2,36 +2,15 @@ import * as yup from 'yup';
 import { Avatar, Button, Label, Modal } from 'flowbite-react';
 import { Field, Form, Formik, FormikHelpers } from 'formik';
 import {
-	IMG_MAX_HEIGHT,
-	IMG_MAX_WIDTH,
 	apiStatusCodes,
-	imageSizeAccepted,
 } from '../../config/CommonConstant';
-import { calculateSize, dataURItoBlob } from '../../utils/CompressImage';
 import React, { useEffect, useState } from 'react';
 import { AlertComponent } from '../AlertComponent';
 import type { AxiosResponse } from 'axios';
 import { updateOrganization } from '../../api/organization';
-import type { Organisation } from './interfaces';
+import type { EditOrgdetailsModalProps, ILogoImage, Organisation, Values } from './interfaces';
 import defaultUserIcon from '../../../public/images/person_FILL1_wght400_GRAD0_opsz24.svg';
-
-interface Values {
-	website: any;
-	name: string;
-	description: string;
-}
-interface ILogoImage {
-	logoFile: string | File;
-	imagePreviewUrl: string | ArrayBuffer | null | File;
-	fileName: string;
-}
-interface EditOrgdetailsModalProps {
-	openModal: boolean;
-	setMessage: (message: string) => void;
-	setOpenModal: (flag: boolean) => void;
-	onEditSucess?: () => void;
-	orgData: Organisation | null;
-}
+import { processImage } from '../../utils/enums/processImage';
 
 const EditOrgdetailsModal = (props: EditOrgdetailsModalProps) => {
 	const [logoImage, setLogoImage] = useState<ILogoImage>({
@@ -85,86 +64,20 @@ const EditOrgdetailsModal = (props: EditOrgdetailsModalProps) => {
 		}
 	}, [props.openModal]);
 
-	const ProcessImg = (e: any): string | undefined => {
-		const file = e?.target.files[0];
-		if (!file) {
-			return;
-		}
-
-		const reader = new FileReader();
-		reader.readAsDataURL(file);
-
-		reader.onload = (event): void => {
-			const imgElement = document.createElement('img');
-			if (imgElement) {
-				imgElement.src =
-					typeof event?.target?.result === 'string' ? event.target.result : '';
-				imgElement.onload = (e): void => {
-					let fileUpdated: File | string = file;
-					let srcEncoded = '';
-					const canvas = document.createElement('canvas');
-
-					const { width, height, ev } = calculateSize(
-						imgElement,
-						IMG_MAX_WIDTH,
-						IMG_MAX_HEIGHT,
-					);
-					canvas.width = width;
-					canvas.height = height;
-
-					const ctx = canvas.getContext('2d');
-					if (ctx && e?.target) {
-						ctx.imageSmoothingEnabled = true;
-						ctx.imageSmoothingQuality = 'high';
-						ctx.drawImage(ev, 0, 0, canvas.width, canvas.height);
-						srcEncoded = ctx.canvas.toDataURL(ev, file.type);
-						const blob = dataURItoBlob(srcEncoded, file.type);
-						fileUpdated = new File([blob], file.name, {
-							type: file.type,
-							lastModified: new Date().getTime(),
-						});
-						setLogoImage({
-							logoFile: fileUpdated,
-							imagePreviewUrl: srcEncoded,
-							fileName: file.name,
-						});
-					}
-				};
-			}
-		};
-	};
-
-	const isEmpty = (object: any): boolean => {
-		for (const property in object) {
-			return false;
-		}
-		return true;
-	};
 	const handleImageChange = (event: any): void => {
-		setImgError('');
-		const reader = new FileReader();
-		const file = event?.target?.files;
-
-		const fieSize = Number((file[0]?.size / 1024 / 1024)?.toFixed(2));
-		const extension = file[0]?.name
-			?.substring(file[0]?.name?.lastIndexOf('.') + 1)
-			?.toLowerCase();
-		if (extension === 'png' || extension === 'jpeg' || extension === 'jpg') {
-			if (fieSize <= imageSizeAccepted) {
-				reader.onloadend = (): void => {
-					ProcessImg(event);
-					isEmpty(reader.result);
-				};
-				reader.readAsDataURL(file[0]);
-				event.preventDefault();
-			} else {
-				setImgError('Please check image size');
-			}
-		} else {
-			setImgError('Invalid image type');
-		}
-	};
-
+    setImgError('');
+    processImage(event, (result, error) => {
+      if (result) {
+        setLogoImage({
+          logoFile: '',
+          imagePreviewUrl: result,
+          fileName: event.target.files[0].name,
+        });
+      } else {
+        setImgError(error || 'An error occurred while processing the image.');
+      }
+    });
+  };
 	const submitUpdateOrganization = async (values: Values) => {
 		setLoading(true);
 
@@ -254,11 +167,10 @@ const EditOrgdetailsModal = (props: EditOrgdetailsModalProps) => {
 						<Form className="space-y-6" onSubmit={formikHandlers.handleSubmit}>
 							<div className="mb-4 bg-white border border-gray-200 rounded-lg shadow-sm 2xl:col-span-2 dark:border-gray-700 sm:p-6 dark:bg-gray-800">
 								<div className="flex flex-col items-center sm:flex-row 2xl:flex-row p-2 gap-0 sm:gap-4">
-									{typeof logoImage.logoFile === 'string' &&
-									props?.orgData?.logoUrl ? (
+									{logoImage?.imagePreviewUrl ? (
 										<img
 											className="mb-4 rounded-lg w-28 h-28 sm:mb-0 xl:mb-4 2xl:mb-0"
-											src={props?.orgData?.logoUrl}
+											src={logoImage?.imagePreviewUrl}
 											alt="Jese picture"
 										/>
 									) : typeof logoImage.logoFile === 'string' ? (
