+++ src/components/organization/interfaces/index.ts
@@ -167,4 +167,20 @@ export interface IOrgData {
 export interface IExploreOrg {
     orgData: IOrgData
 }
-
+export interface Values {
+	website: any;
+	name: string;
+	description: string;
+}
+export interface ILogoImage {
+	logoFile: string | File;
+	imagePreviewUrl: string | ArrayBuffer | null | File;
+	fileName: string;
+}
+export interface EditOrgdetailsModalProps {
+	openModal: boolean;
+	setMessage: (message: string) => void;
+	setOpenModal: (flag: boolean) => void;
+	onEditSucess?: () => void;
+	orgData: Organisation | null;
+}
